code review rework
