+++ src/api/routers/participants/participantsRouter.ts
@@ -2,19 +2,17 @@ import { AxiosError } from 'axios';
 import express, { Response } from 'express';
 import { z } from 'zod';
 
-import { ApiRole, ApiRoleDTO } from '../../entities/ApiRole';
+import { ApiRoleDTO } from '../../entities/ApiRole';
 import { AuditAction } from '../../entities/AuditTrail';
 import {
-  AdminParticipantCreationPartial,
   Participant,
   ParticipantApprovalPartial,
-  participantCreationAndApprovalPartial,
   ParticipantCreationPartial,
   ParticipantDTO,
   ParticipantStatus,
 } from '../../entities/Participant';
 import { ParticipantType } from '../../entities/ParticipantType';
-import { User, UserCreationPartial, UserDTO, UserRole } from '../../entities/User';
+import { UserDTO, UserRole } from '../../entities/User';
 import { getTraceId } from '../../helpers/loggingHelpers';
 import { mapClientTypeToParticipantType } from '../../helpers/siteConvertingHelpers';
 import { getKcAdminClient } from '../../keycloakAdminClient';
@@ -38,7 +36,6 @@ import {
   validateApiRoles,
 } from '../../services/apiKeyService';
 import {
-  insertAddParticipantAuditTrail,
   insertApproveAccountAuditTrail,
   insertKeyPairAuditTrails,
   insertManageApiKeyAuditTrail,
@@ -70,6 +67,7 @@ import {
   getAllUserFromParticipant,
 } from '../../services/usersService';
 import { createBusinessContactsRouter } from '../businessContactsRouter';
+import { createParticipant } from './participantsCreation';
 import { getParticipantKeyPairs } from './participantsKeyPairs';
 import { getParticipantUsers } from './participantsUsers';
 
@@ -220,70 +218,61 @@ export function createParticipantsRouter() {
     }
   );
 
-  participantsRouter.put('/', isApproverCheck, async (req, res) => {
-    const participantRequest = AdminParticipantCreationPartial.parse(req.body);
-    const existingParticipant = await Participant.query().findOne(
-      'name',
-      participantRequest.participantName
-    );
-    if (existingParticipant) {
-      return res.status(400).send('Duplicate participant name');
-    }
-    const existingUser = await findUserByEmail(participantRequest.email);
-    if (existingUser) {
-      return res.status(400).send('Duplicate requesting user');
-    }
-    const kcAdminClient = await getKcAdminClient();
-    const existingKcUser = await kcAdminClient.users.find({ email: participantRequest.email });
-    if (existingKcUser.length > 0) {
-      return res.status(400).send('Requesting user already exists in Keycloak');
-    }
-
-    const traceId = getTraceId(req);
-    const requestingUser = await findUserByEmail(req.auth?.payload?.email as string);
-    const user = UserCreationPartial.parse({
-      ...req.body,
-      acceptedTerms: true,
-    });
-
-    const types = await ParticipantType.query().findByIds(participantRequest.participantTypes);
-    const apiRoles = await ApiRole.query().findByIds(participantRequest.apiRoles);
-
-    const participantData = participantCreationAndApprovalPartial.parse({
-      name: participantRequest.participantName,
-      types,
-      apiRoles,
-      status: ParticipantStatus.Approved,
-      siteId: participantRequest.siteId,
-      users: [user],
-    });
-
-    const auditTrail = await insertAddParticipantAuditTrail(requestingUser!.email, participantData);
-
-    // create site (UID2-2631)
-
-    // create participant, user, and role/type mappings
-    await Participant.query().insertGraphAndFetch([participantData], {
-      relate: true,
-    });
-
-    // Get newly created user
-    const newUser = (await User.query().findOne('email', participantRequest.email)) as User;
-
-    // create keyCloak user
-    const newKcUser = await createNewUser(
-      kcAdminClient,
-      participantRequest.firstName,
-      participantRequest.lastName,
-      participantRequest.email
-    );
-
-    await sendInviteEmail(kcAdminClient, newKcUser);
-    await sendParticipantApprovedEmail([newUser!], traceId);
-    await updateAuditTrailToProceed(auditTrail.id);
-
-    return res.sendStatus(200);
-  });
+  participantsRouter.put('/', createParticipant); // isApproverCheck, async (req, res) => {
+  //   const participantRequest = AdminParticipantCreationPartial.parse(req.body);
+  //   const existingParticipant = await Participant.query().findOne(
+  //     'name',
+  //     participantRequest.participantName
+  //   );
+  //   if (existingParticipant) {
+  //     return res.status(400).send('Duplicate participant name');
+  //   }
+  //   const existingUser = await findUserByEmail(participantRequest.email);
+  //   if (existingUser) {
+  //     return res.status(400).send('Duplicate requesting user');
+  //   }
+  //   const kcAdminClient = await getKcAdminClient();
+  //   const existingKcUser = await kcAdminClient.users.find({ email: participantRequest.email });
+  //   if (existingKcUser.length > 0) {
+  //     return res.status(400).send('Requesting user already exists in Keycloak');
+  //   }
+  //   const traceId = getTraceId(req);
+  //   const requestingUser = await findUserByEmail(req.auth?.payload?.email as string);
+  //   const user = UserCreationPartial.parse({
+  //     ...req.body,
+  //     acceptedTerms: true,
+  //   });
+  //   const types = await ParticipantType.query().findByIds(participantRequest.participantTypes);
+  //   const apiRoles = await ApiRole.query().findByIds(participantRequest.apiRoles);
+  //   const participantData = participantCreationAndApprovalPartial.parse({
+  //     name: participantRequest.participantName,
+  //     types,
+  //     apiRoles,
+  //     status: ParticipantStatus.Approved,
+  //     siteId: participantRequest.siteId,
+  //     users: [user],
+  //   });
+  //   const auditTrail = await insertAddParticipantAuditTrail(requestingUser!.email, participantData);
+  //   // create site (UID2-2631)
+  //   // create participant, user, and role/type mappings
+  //   await Participant.query().insertGraphAndFetch([participantData], {
+  //     relate: true,
+  //   });
+  //   // Get newly created user
+  //   const newUser = (await User.query().findOne('email', participantRequest.email)) as User;
+  //   // create keyCloak user
+  //   const newKcUser = await createNewUser(
+  //     kcAdminClient,
+  //     participantRequest.firstName,
+  //     participantRequest.lastName,
+  //     participantRequest.email
+  //   );
+  //   await sendInviteEmail(kcAdminClient, newKcUser);
+  //   await sendParticipantApprovedEmail([newUser!], traceId);
+  //   await updateAuditTrailToProceed(auditTrail.id);
+  //   return res.sendStatus(200);
+  // });
+  // });
 
   participantsRouter.use('/:participantId', checkParticipantId);
