+++ src/api/services/adminServiceHelpers.ts
@@ -50,24 +50,12 @@ export type KeyPairDTO = {
   name?: string;
 };
 
-export const AllowedSiteRoles: Record<string, AvailableClientRole[]> = {
-  [ParticipantTypeData.Advertiser.typeName]: ['MAPPER', 'SHARER'],
-  [ParticipantTypeData.DataProvider.typeName]: ['MAPPER', 'SHARER'],
-  [ParticipantTypeData.DSP.typeName]: ['ID_READER', 'SHARER'],
-  [ParticipantTypeData.Publisher.typeName]: ['GENERATOR', 'SHARER'],
-};
-
 export const AllowedSiteRolesById: Record<number, number[]> = {
-  1: [3],
-  2: [1],
-  3: [1],
-  4: [2],
+  1: [3], // DSP: ID_READER
+  2: [1], // Advertiser: MAPPER
+  3: [1], // Data Provider: MAPPER
+  4: [2], // Publisher: GENERATOR
 };
-
-export function GetRecommendedRoles(roles: ParticipantTypeDTO[]) {
-  const recommendedRolesWithDuplicates = roles.flatMap((r) => AllowedSiteRoles[r.typeName]);
-  return [...new Set(recommendedRolesWithDuplicates)];
-}
 
 export function GetRecommendedRolesById(roleIds: number[]) {
   const recommendedRolesWithDuplicates = roleIds.flatMap((r) => AllowedSiteRolesById[r]);
