+++ src/web/components/ParticipantManagement/AddParticipantDialog.tsx
@@ -1,5 +1,5 @@
 import Fuse from 'fuse.js';
-import { ChangeEvent, useEffect, useMemo, useState } from 'react';
+import { ChangeEvent, useCallback, useEffect, useMemo, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
 import { ApiRoleDTO } from '../../../api/entities/ApiRole';
@@ -9,6 +9,7 @@ import { GetRecommendedRolesById, SiteDTO } from '../../../api/services/adminSer
 import { AddParticipantForm } from '../../services/participant';
 import { useSiteList } from '../../services/site';
 import { sortApiRoles } from '../../utils/apiRoles';
+import { extractMessageFromAxiosError } from '../../utils/errorHelpers';
 import { Dialog } from '../Core/Dialog';
 import { CheckboxInput } from '../Input/CheckboxInput';
 import { Input } from '../Input/Input';
@@ -46,8 +47,16 @@ function AddParticipantDialog({
   const [searchText, setSearchText] = useState('');
   const [selectedSite, setSelectedSite] = useState<SiteDTO>();
 
-  const formMethods = useForm<AddParticipantForm>();
-  const { register, setValue, watch, handleSubmit, reset } = formMethods;
+  const formMethods = useForm<AddParticipantForm>({ defaultValues: { siteIdType: 0 } });
+  const {
+    register,
+    setValue,
+    watch,
+    handleSubmit,
+    reset,
+    setError,
+    formState: { errors },
+  } = formMethods;
 
   useEffect(() => {
     setSiteSearchResults(fuse?.search(searchText));
@@ -70,10 +79,30 @@ function AddParticipantDialog({
     }
   }, [open, reset]);
 
-  const onSubmit = async (formData: AddParticipantForm) => {
-    await onAddParticipant(formData);
-    setOpen(false);
-  };
+  const onSubmit = useCallback(
+    async (formData: AddParticipantForm) => {
+      await onAddParticipant(formData);
+      setOpen(false);
+    },
+    [onAddParticipant]
+  );
+
+  const submit = useCallback(
+    async (formData: AddParticipantForm) => {
+      try {
+        await onSubmit(formData);
+      } catch (err) {
+        const message =
+          extractMessageFromAxiosError(err as Error) ?? 'Something went wrong, please try again';
+
+        setError('root.serverError', {
+          type: '400',
+          message,
+        });
+      }
+    },
+    [onSubmit, setError]
+  );
 
   const onSiteClick = (site: SiteDTO) => {
     setValue('siteId', site.id);
@@ -93,19 +122,22 @@ function AddParticipantDialog({
       onOpenChange={setOpen}
       className='add-participant-dialog'
     >
+      {errors.root?.serverError && (
+        <p className='form-error' data-testid='formError'>
+          {errors.root?.serverError.message}
+        </p>
+      )}
       <FormProvider {...formMethods}>
-        <form onSubmit={handleSubmit(onSubmit)}>
+        <form onSubmit={handleSubmit(submit)}>
           <h4>Participant Information</h4>
-          <span>Provide the following information about the company/participant</span>
+          <span>Provide the following information about the company/participant.</span>
           <div className='add-participant-dialog-flex'>
-            <div>
-              <TextInput
-                inputName='participantName'
-                label='Participant Name'
-                className='text-input'
-                rules={{ required: 'Please specify Participant Name.' }}
-              />
-            </div>
+            <TextInput
+              inputName='participantName'
+              label='Participant Name'
+              className='text-input'
+              rules={{ required: 'Please specify Participant Name.' }}
+            />
 
             <div className='right-column'>
               <CheckboxInput
@@ -121,12 +153,12 @@ function AddParticipantDialog({
           </div>
           <div>
             <div className='add-participant-dialog-flex'>
-              <div className='left-column site-type'>
+              <div className='site-type'>
                 <RadioInput
                   inputName='siteIdType'
                   label='Site ID'
                   options={[
-                    { optionLabel: 'Existing Site ID', value: 0, checked: true },
+                    { optionLabel: 'Existing Site ID', value: 0 },
                     { optionLabel: 'New Site ID', value: 1, disabled: true },
                   ]}
                 />
@@ -169,29 +201,16 @@ function AddParticipantDialog({
             </div>
           </div>
           <div className='add-participant-dialog-flex'>
-            <div className='left-column user-roles'>
-              <SelectInput
-                inputName='role'
-                label='Job Function'
-                rules={{ required: 'Please specify your job function.' }}
-                options={(Object.keys(UserRole) as Array<keyof typeof UserRole>).map((key) => ({
-                  optionLabel: UserRole[key],
-                  value: UserRole[key],
-                }))}
-              />
-            </div>
-            <div className='right-column'>
-              <CheckboxInput
-                inputName='apiRoles'
-                label='API Roles'
-                options={sortApiRoles(apiRoles).map((p) => ({
-                  optionLabel: p.externalName,
-                  optionToolTip: p.roleName,
-                  value: p.id,
-                }))}
-                rules={{ required: 'Please specify API Role(s).' }}
-              />
-            </div>
+            <CheckboxInput
+              inputName='apiRoles'
+              label='API Roles'
+              options={sortApiRoles(apiRoles).map((p) => ({
+                optionLabel: p.externalName,
+                optionToolTip: p.roleName,
+                value: p.id,
+              }))}
+              rules={{ required: 'Please specify API Role(s).' }}
+            />
           </div>
           <div>
             <h4>Participant Contact Information</h4>
@@ -207,12 +226,6 @@ function AddParticipantDialog({
                   className='text-input'
                   rules={{ required: 'Please specify Contact First Name.' }}
                 />
-                <TextInput
-                  inputName='email'
-                  label='Contact Email'
-                  className='text-input'
-                  rules={{ required: 'Please specify Contact Email.' }}
-                />
               </div>
               <div className='right-column'>
                 <TextInput
@@ -223,11 +236,34 @@ function AddParticipantDialog({
                 />
               </div>
             </div>
+            <div className='add-participant-dialog-flex'>
+              <div>
+                <TextInput
+                  inputName='email'
+                  label='Contact Email'
+                  className='text-input'
+                  rules={{ required: 'Please specify Contact Email.' }}
+                />
+              </div>
+              <div>
+                <div className='user-roles right-column'>
+                  <SelectInput
+                    inputName='role'
+                    label='Job Function'
+                    rules={{ required: "Please specify the contact's job function." }}
+                    options={(Object.keys(UserRole) as Array<keyof typeof UserRole>).map((key) => ({
+                      optionLabel: UserRole[key],
+                      value: UserRole[key],
+                    }))}
+                  />
+                </div>
+              </div>
+            </div>
           </div>
           <br />
           <div className='request-button'>
             <button type='submit' className='primary-button'>
-              Request Account
+              Add Participant
             </button>
           </div>
         </form>
