+++ src/web/screens/manageParticipants.scss
